refactor: a better place for useControlledDialog
