+++ packages/maskbook/src/plugins/Collectible/SNSAdaptor/ActionBar.tsx
@@ -2,7 +2,7 @@ import { makeStyles, Box } from '@material-ui/core'
 import { useI18N } from '../../../utils'
 import { CollectibleState } from '../hooks/useCollectibleState'
 import ActionButton from '../../../extension/options-page/DashboardComponents/ActionButton'
-import { useControlledDialog } from './useControlledDialog'
+import { useControlledDialog } from '../../../utils/hooks/useControlledDialog'
 import { MakeOfferDialog } from './MakeOfferDialog'
 import { PostListingDialog } from './PostListingDialog'
 import { CheckoutDialog } from './CheckoutDialog'
