+++ packages/maskbook/src/utils/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useControlledDialog'
 export * from './useMatchXS'
 export * from './useMenu'
 export * from './useQRCodeImageScan'
